Add files via upload
For the code related to how the convex controller runs use the convex labeled convexMainRun. convexOrbitsV4, and convexPlots. Note the most up to date convex controller is on the google colab. For code related to the kp kd controller use the mainRun.py, orbitPlots.py, and orbits.py. The most up to date stan files are the pdControllerV2.rmd and the pdControllerV3.stan.